+++ app/src/main/java/com/example/shiftproj/feature/cityTemp/add/presentation/AddCityTempView.kt
@@ -3,5 +3,5 @@ package com.example.shiftproj.feature.cityTemp.add.presentation
 import com.example.shiftproj.feature.util.ViewContract
 
 interface AddCityTempView : ViewContract {
-    fun onAddButtonClick(city : String, temp : Double)
+    fun onAddButtonClick(city : String, temp : Long)
 }
